# api-moodle/openapi.yaml
openapi: 3.0.3
info:
  title: API Moodle — UOR Connect
  version: 0.2.0
  description: |
    Contrato interno da camada de integração Moodle → UOR Connect.

    A API expõe somente dados pedagógicos normalizados. O cliente nunca recebe
    cookies, `sesskey`, credenciais, HTML, URLs ou identificadores internos do
    Moodle. Notas oficiais, matrículas e dados financeiros continuam a pertencer
    ao Sistema da Secretaria.

    Todas as operações desta versão pertencem ao MVP. Enquanto o handler real,
    o gateway Moodle, a validação de ownership e os testes de contrato não
    estiverem ligados em conjunto, a operação permanece marcada como `planned`.
  contact:
    name: Equipa UOR Connect
servers:
  - url: http://127.0.0.1:3333
    description: Backend local direto
  - url: http://127.0.0.1:8082/api
    description: Gateway do frontend local
  - url: https://api.uorconnect.ao
    description: API de produção direta
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
  /integrations/moodle/session:
    post:
      tags: [Sessão Moodle]
      summary: Ligar ou atualizar a conta Moodle
      description: |
        Confirma a identidade Moodle, cifra credenciais e sessão separadamente e
        agenda a sincronização inicial. `rememberCredentials` deve ser `true`.

        Com autenticação UOR Connect por cookie, envie também `X-CSRF-Token`. O
        Swagger recomenda Bearer porque não tenta obter o cookie CSRF da aplicação.
      operationId: createMoodleSession
      x-implementation-status: implemented
      x-source-status: observed
      x-phase: mvp
      x-cache-ttl: 0
      x-read-only: false
      x-requires-moodle-admin: false
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Credenciais de uma ligação existente atualizadas.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
              example:
                data:
                  connection:
                    status: CONNECTED
                    connected: true
                    credentialsStored: true
                    actionRequired: none
                    retryable: false
                    lastAuthenticatedAt: '2026-07-19T12:00:00.000Z'
                    lastSuccessfulSyncAt: '2026-07-19T11:45:00.000Z'
                  initialSyncRunId: 550e8400-e29b-41d4-a716-446655440010
                meta:
                  requestId: req_01J2Q6W7A8B9C0D1E2F3G4H5J6
                  syncedAt: '2026-07-19T11:45:00.000Z'
                  stale: false
        '201':
          description: Nova ligação Moodle criada.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
              example:
                data:
                  connection:
                    status: CONNECTED
                    connected: true
                    credentialsStored: true
                    actionRequired: none
                    retryable: false
                    lastAuthenticatedAt: '2026-07-19T12:00:00.000Z'
                    lastSuccessfulSyncAt: null
                  initialSyncRunId: 550e8400-e29b-41d4-a716-446655440010
                meta:
                  requestId: req_01J2Q6W7A8B9C0D1E2F3G4H5J6
                  syncedAt: null
                  stale: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedJsonMediaType'
        '422':
          $ref: '#/components/responses/SessionUnprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/UpstreamInvalid'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Sessão Moodle]
      summary: Desligar a conta Moodle
      description: |
        Operação idempotente. Invalida a geração da ligação e elimina localmente
        credenciais, sessão, snapshots e sincronizações, mesmo se o logout Moodle
        em best effort falhar.
      operationId: deleteMoodleSession
      x-implementation-status: implemented
      x-source-status: observed
      x-phase: mvp
      x-cache-ttl: 0
      x-read-only: false
      x-requires-moodle-admin: false
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Conta desligada ou já desligada.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
              example:
                data:
                  connection:
                    status: DISCONNECTED
                    connected: false
                    credentialsStored: false
                    actionRequired: connect
                    retryable: false
                    lastAuthenticatedAt: null
                    lastSuccessfulSyncAt: null
                  initialSyncRunId: null
                meta:
                  requestId: req_01J2Q75Z0A1B2C3D4E5F6G7H8J
                  syncedAt: null
                  stale: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /integrations/moodle/me:
    get:
      tags: [Visão geral]
      summary: Obter a identidade Moodle normalizada
      operationId: getMoodleProfile
      x-implementation-status: implemented
      x-source-status: observed
      x-phase: mvp
      x-cache-ttl: 600
      x-read-only: true
      x-requires-moodle-admin: false
      responses:
        '200':
          description: Identidade mínima da conta ligada.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
              example:
                data:
                  id: 550e8400-e29b-41d4-a716-446655440001
                  studentNumber: '20260001'
                  displayName: Estudante UOR
                  email: estudante@example.test
                  timezone: Africa/Luanda
                  lastSyncedAt: '2026-07-19T12:05:00.000Z'
                meta:
                  requestId: req_01J2Q7E5K6M7N8P9Q0R1S2T3V4
                  syncedAt: '2026-07-19T12:05:00.000Z'
                  stale: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/ConnectionRequired'
        '502':
          $ref: '#/components/responses/UpstreamInvalid'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /integrations/moodle/overview:
    get:
      tags: [Visão geral]
      summary: Obter o resumo pedagógico Moodle
      description: |
        Retorna totais com qualidade explícita, cobertura do snapshot e média
        calculada apenas sobre disciplinas cujo acompanhamento está configurado.
      operationId: getMoodleOverview
      x-implementation-status: implemented
      x-source-status: derived
      x-phase: mvp
      x-cache-ttl: 600
      x-read-only: true
      x-requires-moodle-admin: false
      responses:
        '200':
          description: Resumo útil do snapshot atual.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewEnvelope'
              examples:
                complete:
                  summary: Disciplinas completas e materiais ainda parciais
                  value:
                    data:
                      connection:
                        status: CONNECTED
                        connected: true
                        credentialsStored: true
                        actionRequired: none
                        retryable: false
                        lastAuthenticatedAt: '2026-07-19T12:00:00.000Z'
                        lastSuccessfulSyncAt: '2026-07-19T12:05:00.000Z'
                      counts:
                        courses:
                          value: 29
                          status: exact
                        coursesVisible:
                          value: 29
                          status: exact
                        coursesWithProgress:
                          value: 5
                          status: exact
                        materials:
                          value: 32
                          status: partial
                        activities:
                          value: null
                          status: unsupported
                        assignmentsOpen:
                          value: null
                          status: unsupported
                        quizzesOpen:
                          value: null
                          status: unsupported
                        notificationsUnread:
                          value: null
                          status: unsupported
                      coverage:
                        processedCourses: 5
                        totalCourses: 29
                        failedCourses: 0
                      progress:
                        status: partial
                        trackedCourses: 5
                        untrackedCourses: 24
                        averagePercent: 36.4
                      nextDeadlines:
                        status: unsupported
                        items: []
                    meta:
                      requestId: req_01J2Q7J5K6M7N8P9Q0R1S2T3V4
                      syncedAt: '2026-07-19T12:05:00.000Z'
                      stale: false
                      snapshotVersion: 7
                notSynced:
                  summary: Ligação criada, mas o primeiro snapshot ainda não terminou
                  value:
                    data:
                      connection:
                        status: CONNECTED
                        connected: true
                        credentialsStored: true
                        actionRequired: none
                        retryable: false
                        lastAuthenticatedAt: '2026-07-19T12:00:00.000Z'
                        lastSuccessfulSyncAt: null
                      counts:
                        courses:
                          value: null
                          status: not_synced
                        coursesVisible:
                          value: null
                          status: not_synced
                        coursesWithProgress:
                          value: null
                          status: not_synced
                        materials:
                          value: null
                          status: not_synced
                        activities:
                          value: null
                          status: unsupported
                        assignmentsOpen:
                          value: null
                          status: unsupported
                        quizzesOpen:
                          value: null
                          status: unsupported
                        notificationsUnread:
                          value: null
                          status: unsupported
                      coverage:
                        processedCourses: 0
                        totalCourses: null
                        failedCourses: 0
                      progress:
                        status: not_synced
                        trackedCourses: null
                        untrackedCourses: null
                        averagePercent: null
                      nextDeadlines:
                        status: unsupported
                        items: []
                    meta:
                      requestId: req_01J2Q7K5K6M7N8P9Q0R1S2T3V4
                      syncedAt: null
                      stale: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/ConnectionRequired'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /integrations/moodle/courses:
    get:
      tags: [Disciplinas]
      summary: Listar disciplinas sincronizadas
      operationId: listMoodleCourses
      x-implementation-status: implemented
      x-source-status: observed
      x-phase: mvp
      x-cache-ttl: 600
      x-read-only: true
      x-requires-moodle-admin: false
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Página estável de disciplinas do snapshot atual.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListEnvelope'
              example:
                data:
                  - id: 550e8400-e29b-41d4-a716-446655440002
                    name: Engenharia de Software
                    shortName: INF-301
                    category: Engenharia Informática
                    description: Conteúdos e atividades da disciplina.
                    startDate: '2026-02-02T00:00:00.000Z'
                    endDate: null
                    visible: true
                    favourite: true
                    progressAvailable: true
                    progressPercent: 42.5
                    stale: false
                    lastSyncedAt: '2026-07-19T12:05:00.000Z'
                  - id: 550e8400-e29b-41d4-a716-446655440003
                    name: Matemática Discreta
                    shortName: MAT-201
                    category: Ciências Exatas
                    description: null
                    startDate: '2026-02-02T00:00:00.000Z'
                    endDate: null
                    visible: true
                    favourite: false
                    progressAvailable: false
                    progressPercent: null
                    stale: false
                    lastSyncedAt: '2026-07-19T12:05:00.000Z'
                meta:
                  requestId: req_01J2Q84K6M7N8P9Q0R1S2T3V4W
                  syncedAt: '2026-07-19T12:05:00.000Z'
                  stale: false
                  snapshotVersion: 7
                  pagination:
                    returned: 2
                    limit: 20
                    hasMore: true
                    nextCursor: cur_v1_7M9K2N4P6R8T
                    total: 29
                    totalStatus: exact
                  coverage:
                    processedCourses: 29
                    totalCourses: 29
                    failedCourses: 0
        '400':
          $ref: '#/components/responses/InvalidCursor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/SnapshotChanged'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT UOR Connect. Recomendado para o Swagger Try it out.
    cookieAuth:
      type: apiKey
      in: cookie
      name: uor_auth
      description: |
        Cookie de sessão UOR Connect. Mutações também exigem o header CSRF da
        aplicação; a Swagger UI não obtém esse valor automaticamente.
  headers:
    PrivateNoStore:
      description: Respostas pessoais não podem ser guardadas por caches partilhados.
      schema:
        type: string
        example: private, no-store
...
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Obrigatório apenas quando a autenticação UOR usa cookie.
      schema:
        type: string
        minLength: 16
...
  responses:
    Unauthorized:
      description: JWT ou cookie UOR Connect ausente ou inválido.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UOR_AUTH_REQUIRED
              message: Autenticação UOR Connect necessária.
              retryable: false
              actionRequired: none
            meta:
              requestId: req_01J2Q9AK6M7N8P9Q0R1S2T3V4W
...
  schemas:
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [username, password, rememberCredentials]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 128
          writeOnly: true
          description: Deve ser o próprio studentNumber UOR autenticado. Divergências são rejeitadas localmente antes de contactar o Moodle; o valor nunca é devolvido nem registado.
        password:
          type: string
          minLength: 1
          maxLength: 512
          format: password
          writeOnly: true
          description: Palavra-passe Moodle; nunca é devolvida nem registada.
        rememberCredentials:
          type: boolean
          enum: [true]
          description: Consentimento obrigatório para renovação automática cifrada.
...
    SessionEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/SessionData'
        meta:
          $ref: '#/components/schemas/BasicMeta'
    ProfileEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Profile'
        meta:
          $ref: '#/components/schemas/BasicMeta'
...